+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Internal**:
 
 - Have relay generate metric billing outcomes. ([#6066](https://github.com/getsentry/relay/pull/6066))
+- Update sentry-conventions to 0.12.0.
 
 ## 26.6.0